- Updated all versions to 3.0.0-SNAPSHOT - Updated all Flex dependencies to Apache Flex 4.15.0 - Updated all Air dependencies to 20.0 - Updated all Flash dependencies to 20.0 - Updated to Flexmojos 7.1.0
Got a first version that is able to compile, if I exclude the compilation of all tests by adding "-Dmaven.test.skip=true".

Currently the build relies greatly on a number of maven artifacts which are not publicly available in maven repositories. We have to figure out how to fix this.

I'll concentrate on the compilation of tests next.